fix(cookie-consent): adjust draft state in render, rename provider, split null test (#2340)
