+++ apps/frontend/src/components/cookie-consent/use-consent.ts
@@ -1,4 +1,4 @@
-import { CookieConsentContext } from '@/components/cookie-consent/cookie-consent-provider';
+import { CookieConsentContext } from '@/components/cookie-consent/CookieConsentProvider';
 import { useContext } from 'react';
 
 export const useConsent = () => {
